refactor: moved related utils into SingleOrRangeNumberCondition (M2-6704)
